fixed #27
also simplified handling of shorthand notations like +=
writing x+=2 is now only a shortcut for writing x=x+1.
This changes the meaning of the statement a bit because the
expression x is now evaluated twice instead of only once.
however for most use cases this should make no difference and
so I thought it was justified to simplify this.