added option to remove dispatch-checks
this option removes the null-checks and the checks for accessing
destroyed objects. It is interesting to see how this can reduce the file
size, but I would not recommend using this option.
If people have performance problems I could still add a feature to
disable the checks for some performance critical parts of the code.