add damage type implementation
This is an implementation of DamageType, a system designed by
looking_for_help which uses the runed bracer ability to differentiate
basic attacks from spell damage in the context of a damage event. This
script is a conersion of DamageType by Cokemonkey11, and adjusted to
use Wurst DamageDetection instead of StructuredDD (vJass).