[stl] fix the vec3.polarProject method
this method was previously using airAngle as 0 for vertical and PI/2
for horizontal. Clearly this is the design of a madman. I can see that
(from hudson) the only project currently using polarProject method is
Forest Defense, but it uses airAngle=0, so I amguessing that frotty has
already fixed this bug in STL2.