fixed module constructor calls
the problem was that if an intermediate module had no constructor the
upstream constructors would not get called.

Or if a module had two constructors, then upstream constructors would be
called twice.

No it is enforced that every module has exactly one constructor.
(A default is generated, if no constructor is given)