fixed #737 optimizer bug
The problem was in TempMerger, which remembers the current value of
variables and inlines temporary variables.
It checks that moving expressions has no unintended side-effects.
However the necessary checks (whether a statement reads a global or
calls a function) was cached and not recomputed when statements changed
due to optimizations.

This fix removes the cache, which probably makes this optimization
slower.