fixed global initialization code
for global initialization, we have two places: init functions for
packages and a map that stores the initialization code for each
variable. The map is used to avoid running all init functions when
executing code at compiletime.

The problem was that the rewrite phases for certain expressions would
not correctly update the map and this would result in two different
initializers in the init function and the map. To avoi this problem,
with this commit the map no longer stores the init expression directly,
but instead keeps a reference to the ImSet init statements in the init
function. This should not be rewritten, since it is a simple array or
global assignment, so the whole process should be more stable.