fix reconciling code for editor
previously, we checked only the direct dependencies of changed files
(actually, it was two levels of dependencies, due to some redundt code).
However, this is only correct, as long as we do not consider classes and modules.
Extending a class or using a module can re-export the members of the
class or module, so not only packages that directly import a changed
package can be affected by a change.

We now had two ways to fix this:
1) fine grained dependency analysis to exactly find other affected files
2) include all transitive dependencies.

Option 1 would have been more efficient, but there is a great chance for
more subtle bugs. So this commit implements option 2, which will make
updates more compute intense, but should be more stable.